+++ R/zzz.R
@@ -3,7 +3,10 @@
     packageStartupMessage(
         c(
             "Currently, the default options are:",
-            paste("\n- location for datasets is", getOption("IssueTrackeR.dataset.dir")),
+            paste(
+                "\n- location for datasets is",
+                getOption("IssueTrackeR.dataset.dir")
+            ),
             paste("\n- owner:", getOption("IssueTrackeR.owner")),
             paste("\n- repo:", getOption("IssueTrackeR.repo"))
         )