+++ NEWS.md
@@ -8,10 +8,15 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+
+* New class `MilestonesTB` for milestones objects
+
 ### Changed
 
 * Removed progress bar in `gh::gh()` calls
 * `update_database()` updates all the milestones (closed and open)
+* `write_issues_to_dataset()`, `write_labels_to_dataset()` and `write_milestones_to_dataset()` are regrouped into `write_to_dataset()`
 
 ### Fixed