+++ NEWS.md
@@ -5,11 +5,32 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-
 ## [Unreleased]
 
-## [1.3.1] - 2025-10-27
+## [1.4.0] - 2026-06-21
 
+### Added
+
+* New class `MilestonesTB` for milestones objects
+* Number of issues open or closed in each milestones
+* New function to plot issues
+* New checked case if internet connection is down
+* Display of the default options at the start of the package [#7](https://github.com/TanguyBarthelemy/IssueTrackeR/issues/7)
+* New function `reset_options` to reset the default options [#4](https://github.com/TanguyBarthelemy/IssueTrackeR/issues/4)
+* Argument `with_labels` display the labels with the issue in a summary call of a list of issues
+
+### Changed
+
+* Removed progress bar in `gh::gh()` calls
+* `update_database()` updates all the milestones (closed and open)
+* `write_issues_to_dataset()`, `write_labels_to_dataset()` and `write_milestones_to_dataset()` are regrouped into `write_to_dataset()`
+* `get_issues()` accept now several repos to look for issues. Argument `repo` can be a vector of several repos.
+
+### Fixed
+
+* Removed a warning due to the use of `ignore.case = TRUE` and `fixed = TRUE` simultaneously in `grepl()`
+
+## [1.3.1] - 2025-10-27
 
 ### Added
 
@@ -18,7 +39,6 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 * `get_nbr_comments()`: Function to compute the number of comments
 * `with_comments()`: Function to filter the issue with or without comments
 * `author_last_comment()`: Add new function to retrieve the name of the last commentator
-
 
 ## [1.3.0] - 2025-10-21
 
@@ -42,7 +62,6 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 * `get_labels()` now accepts `repo = NULL` and retrieve all the labels from all the repos from an owner.
 * slightly change the writing and reading method of datasets
 
-
 ## [1.2.0] - 2025-07-16
 
 ### Changed
@@ -58,7 +77,6 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 * Url link with Issues printed
 * New `sample()`, `rbind()`, `summary()` method
 
-
 ## [1.1.1] - 2025-04-26
 
 ### Added
@@ -71,7 +89,6 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 * `ignore.case = FALSE` if `fixed = TRUE`
 * lint condition_call (with `call. = FALSE`)
-
 
 ## [1.1.0] - 2025-01-09
 
@@ -93,7 +110,6 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 * `[.IssuesTB` was duplicated
 * removed `... = _` in paste0 for R version before 4.2
 
-
 ## [1.0.0] - 2024-09-12
 
 ### Added
@@ -102,7 +118,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 * New CHANGELOG (`NEWS.md`)
 * Documentation for `logic_reducer()`, `no_milestones()`, `vgrepl()` and `simple_sort`
 
-[Unreleased]: https://github.com/TanguyBarthelemy/IssueTrackeR/compare/v1.3.1...HEAD
+[Unreleased]: https://github.com/TanguyBarthelemy/IssueTrackeR/compare/1.4.0...HEAD
+[1.4.0]: https://github.com/TanguyBarthelemy/IssueTrackeR/compare/v1.3.1...1.4.0
 [1.3.1]: https://github.com/TanguyBarthelemy/IssueTrackeR/compare/v1.3.0...v1.3.1
 [1.3.0]: https://github.com/TanguyBarthelemy/IssueTrackeR/compare/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/TanguyBarthelemy/IssueTrackeR/compare/v1.1.1...v1.2.0